3rd Eye for Blind person
I have designed a special wearable device based on the Arduino board which can be worn like a cloth for blinds. This device is equipped with five ultrasonic sensors, consisting of five modules which are connected to the different parts of the body. Among them, two for both shoulder, another two for both knees and one for the hand. Using the five ultrasonic sensors, blind people can detect the objects in a five dimensional view around them and can easily travel anywhere.